pyproject.toml
[tool.poetry]
name = "gar-loader"
version = "0.1.0"
description = ""
authors = ["Feosen <iva588@yandex.ru>"]
readme = "README.md"
packages = [{include = "gar_loader"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
rarfile = "^4.0"
six = "^1.16.0"
zeep = "^4.2.1"
progress = "^1.6"
psycopg2 = "^2.9.6"
suds = "^1.1.2"


[tool.poetry.group.dev.dependencies]
lxml-stubs = "^0.4.0"
mypy = "<1.2"
django-stubs = {extras = ["compatible-mypy"], version = "^1.16.0"}
types-requests = "^2.28.11.17"
black = "^23.3.0"
flake8 = "^6.0.0"
flake8-pyproject = "^1.2.3"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.flake8]
ignore = ['W503']
max-line-length = 120
exclude = [
    '.git',
    '.mypy_cache',
    '.venv',
]

[tool.mypy]
strict = true
ignore_missing_imports = true
disallow_untyped_decorators = false
plugins = [
    'mypy_django_plugin.main',
]
modules = [
    'gar_loader',
    'fias',
    'target',
]

[[tool.mypy.overrides]]
module = [
    'gar_loader',
]
exclude = [
    'settings\.py',
]
[tool.django-stubs]
django_settings_module = 'gar_loader.settings'